+++ src/client/_compiled_header.js
@@ -1,0 +1,10 @@
+// Local Client Behaviors
+// ======================
+// pfraze 2013
+
+if (typeof this.local == 'undefined')
+	this.local = {};
+if (typeof this.local.client == 'undefined')
+	this.local.client = {};
+
+(function() {