+++ src/http/helpers.js
@@ -1,0 +1,114 @@
+// Helpers
+// =======
+
+// parseLinkHeader()
+// EXPORTED
+// breaks a link header into a javascript object
+local.http.parseLinkHeader = function parseLinkHeader(headerStr) {
+	if (typeof headerStr !== 'string') {
+		return headerStr;
+	}
+	// '</foo/bar>; rel="baz"; title="blah", </foo/bar>; rel="baz"; title="blah", </foo/bar>; rel="baz"; title="blah"'
+	return headerStr.replace(/,[\s]*</g, '|||<').split('|||').map(function(linkStr) {
+		// ['</foo/bar>; rel="baz"; title="blah"', '</foo/bar>; rel="baz"; title="blah"']
+		var link = {};
+		linkStr.trim().split(';').forEach(function(attrStr) {
+			// ['</foo/bar>', 'rel="baz"', 'title="blah"']
+			attrStr = attrStr.trim();
+			if (!attrStr) { return; }
+			if (attrStr.charAt(0) === '<') {
+				// '</foo/bar>'
+				link.href = attrStr.trim().slice(1, -1);
+			} else {
+				var attrParts = attrStr.split('=');
+				// ['rel', '"baz"']
+				var k = attrParts[0].trim();
+				var v = attrParts[1].trim().slice(1, -1);
+				link[k] = v;
+			}
+		});
+		return link;
+	});
+};
+
+// EXPORTED
+// looks up a link in the cache and generates the URI
+//  - first looks for a matching rel and title
+//    eg lookupLink(links, 'item', 'foobar'), Link: <http://example.com/some/foobar>; rel="item"; title="foobar" -> http://example.com/some/foobar
+//  - then looks for a matching rel with no title and uses that to generate the link
+//    eg lookupLink(links, 'item', 'foobar'), Link: <http://example.com/some/{title}>; rel="item" -> http://example.com/some/foobar
+local.http.lookupLink = function lookupLink(links, rel, title) {
+	var len = links ? links.length : 0;
+	if (!len) { return null; }
+
+	title = title.toLowerCase();
+
+	// try to find the link with a title equal to the param we were given
+	var match = null;
+	for (var i=0; i < len; i++) {
+		var link = links[i];
+		if (!link) { continue; }
+		// find all links with a matching rel
+		if (link.rel && link.rel.indexOf(rel) !== -1) {
+			// look for a title match to the primary parameter
+			if (link.title) {
+				if (link.title.toLowerCase() === title) {
+					match = link;
+					break;
+				}
+			} else {
+				// no title attribute -- it's the template URI, so hold onto it
+				match = link;
+			}
+		}
+	}
+
+	return match ? match.href : null;
+};
+
+// EXPORTED
+// correctly joins together to url segments
+local.http.joinUrl = function joinUrl() {
+	var parts = Array.prototype.map.call(arguments, function(arg) {
+		var lo = 0, hi = arg.length;
+		if (arg.charAt(0) === '/')      { lo += 1; }
+		if (arg.charAt(hi - 1) === '/') { hi -= 1; }
+		return arg.substring(lo, hi);
+	});
+	return parts.join('/');
+};
+
+// EXPORTED
+// parseUri 1.2.2, (c) Steven Levithan <stevenlevithan.com>, MIT License
+local.http.parseUri = function parseUri(str) {
+	if (typeof str === 'object') {
+		if (str.url) { str = str.url; }
+		else if (str.host || str.path) { str = local.http.joinUrl(req.host, req.path); }
+	}
+	var	o   = local.http.parseUri.options,
+		m   = o.parser[o.strictMode ? "strict" : "loose"].exec(str),
+		uri = {},
+		i   = 14;
+
+	while (i--) uri[o.key[i]] = m[i] || "";
+
+	uri[o.q.name] = {};
+	uri[o.key[12]].replace(o.q.parser, function ($0, $1, $2) {
+		if ($1) uri[o.q.name][$1] = $2;
+	});
+
+	return uri;
+};
+
+local.http.parseUri.options = {
+	strictMode: false,
+	key: ["source","protocol","authority","userInfo","user","password","host","port","relative","path","directory","file","query","anchor"],
+	q:   {
+		name:   "queryKey",
+		parser: /(?:^|&)([^&=]*)=?([^&]*)/g
+	},
+	parser: {
+		strict: /^(?:([^:\/?#]+):)?(?:\/\/((?:(([^:@]*)(?::([^:@]*))?)?@)?([^:\/?#]*)(?::(\d*))?))?((((?:[^?#\/]*\/)*)([^?#]*))(?:\?([^#]*))?(?:#(.*))?)/,
+		loose:  /^(?:(?![^:@]+:[^:@\/]*@)([^:\/?#.]+):)?(?:\/\/)?((?:(([^:@]*)(?::([^:@]*))?)?@)?([^:\/?#]*)(?::(\d*))?)(((\/(?:[^?#](?![^?#\/]*\.[^?#\/.]+(?:[?#]|$)))*\/?)?([^?#\/]*))(?:\?([^#]*))?(?:#(.*))?)/
+	}
+};